+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemStatus/proposalDataListItemStatus.tsx
@@ -8,37 +8,21 @@ import {
     Tag,
     formatterUtils,
     type StatePingAnimationVariant,
-    type TagVariant,
 } from '../../../../../core';
 import type { ModulesCopy } from '../../../../assets';
 import { useOdsModulesContext } from '../../../odsModulesProvider';
-import { type IProposalDataListItemStructureProps, type ProposalStatus } from '../proposalDataListItemStructure';
+import { ProposalStatus, proposalStatusToTagVariant } from '../../proposalUtils';
+import { type IProposalDataListItemStructureProps } from '../proposalDataListItemStructure';
+import { proposalDataListItemUtils } from '../proposalDataListItemUtils';
 
 export interface IProposalDataListItemStatusProps
     extends Pick<IProposalDataListItemStructureProps, 'date' | 'status' | 'voted'> {}
 
-const statusToTagVariant: Record<ProposalStatus, TagVariant> = {
-    accepted: 'success',
-    active: 'info',
-    challenged: 'warning',
-    draft: 'neutral',
-    executed: 'success',
-    expired: 'critical',
-    failed: 'critical',
-    partiallyExecuted: 'warning',
-    pending: 'neutral',
-    queued: 'success',
-    rejected: 'critical',
-    vetoed: 'warning',
-};
-
-type OngoingProposalStatus = 'active' | 'challenged' | 'vetoed';
-
-const ongoingStatusToPingVariant: Record<OngoingProposalStatus, StatePingAnimationVariant> = {
-    active: 'info',
-    challenged: 'warning',
-    vetoed: 'warning',
-};
+const proposalStatusToPingVariant: Map<ProposalStatus, StatePingAnimationVariant> = new Map([
+    [ProposalStatus.ACTIVE, 'info'],
+    [ProposalStatus.CHALLENGED, 'warning'],
+    [ProposalStatus.VETOED, 'warning'],
+]);
 
 const getFormattedProposalDate = (date: string | number, now: number, copy: ModulesCopy) => {
     const formattedDuration = formatterUtils.formatDate(date, { format: DateFormat.DURATION });
@@ -52,32 +36,37 @@ const getFormattedProposalDate = (date: string | number, now: number, copy: Modu
 export const ProposalDataListItemStatus: React.FC<IProposalDataListItemStatusProps> = (props) => {
     const { date, status, voted } = props;
 
-    const ongoing = status === 'active' || status === 'challenged' || status === 'vetoed';
-    const ongoingAndVoted = ongoing && voted;
-    const showStatusMetadata = status !== 'draft';
+    const isOngoing = proposalDataListItemUtils.isOngoingStatus(status);
+    const isOngoingAndVoted = isOngoing && voted;
+    const showStatusMetadata = status !== ProposalStatus.DRAFT;
 
     const { copy } = useOdsModulesContext();
 
     return (
         <div className="flex items-center gap-x-4 md:gap-x-6">
-            <Tag label={status} variant={statusToTagVariant[status]} className="shrink-0 capitalize" />
+            <Tag
+                label={copy.proposalDataListItemStatus.statusLabel[status]}
+                variant={proposalStatusToTagVariant[status]}
+                className="shrink-0"
+            />
             {showStatusMetadata && (
                 <div className="flex flex-1 items-center justify-end gap-x-2 md:gap-x-3">
                     <span
                         className={classNames('text-sm leading-tight md:text-base', {
-                            'text-info-800': status === 'active',
-                            'text-warning-800': status === 'challenged' || status === 'vetoed',
-                            'text-neutral-800': ongoing === false,
+                            'text-info-800': status === ProposalStatus.ACTIVE,
+                            'text-warning-800':
+                                status === ProposalStatus.CHALLENGED || status === ProposalStatus.VETOED,
+                            'text-neutral-800': isOngoing === false,
                         })}
                     >
-                        {ongoingAndVoted && copy.proposalDataListItemStatus.voted}
-                        {!ongoingAndVoted && date != null && (
+                        {isOngoingAndVoted && copy.proposalDataListItemStatus.voted}
+                        {!isOngoingAndVoted && date != null && (
                             <Rerender>{(now) => getFormattedProposalDate(date, now, copy)}</Rerender>
                         )}
                     </span>
-                    {ongoingAndVoted && <AvatarIcon icon={IconType.CHECKMARK} responsiveSize={{ md: 'md' }} />}
-                    {ongoing && !voted && <StatePingAnimation variant={ongoingStatusToPingVariant[status]} />}
-                    {!ongoing && !voted && date && (
+                    {isOngoingAndVoted && <AvatarIcon icon={IconType.CHECKMARK} responsiveSize={{ md: 'md' }} />}
+                    {isOngoing && !voted && <StatePingAnimation variant={proposalStatusToPingVariant.get(status)} />}
+                    {!isOngoing && !voted && date && (
                         <AvatarIcon icon={IconType.CALENDAR} responsiveSize={{ md: 'md' }} />
                     )}
                 </div>