+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemUtils.test.ts
@@ -1,0 +1,18 @@
+import { ProposalStatus } from '../proposalUtils';
+import { proposalDataListItemUtils } from './proposalDataListItemUtils';
+
+describe('proposalDataListItem utils', () => {
+    describe('isOngoingStatus', () => {
+        it('returns true for ongoing proposal statuses', () => {
+            expect(proposalDataListItemUtils.isOngoingStatus(ProposalStatus.ACTIVE)).toBeTruthy();
+            expect(proposalDataListItemUtils.isOngoingStatus(ProposalStatus.CHALLENGED)).toBeTruthy();
+            expect(proposalDataListItemUtils.isOngoingStatus(ProposalStatus.VETOED)).toBeTruthy();
+        });
+
+        it('returns false for non-ongoing proposal statuses', () => {
+            expect(proposalDataListItemUtils.isOngoingStatus(ProposalStatus.DRAFT)).toBeFalsy();
+            expect(proposalDataListItemUtils.isOngoingStatus(ProposalStatus.EXECUTABLE)).toBeFalsy();
+            expect(proposalDataListItemUtils.isOngoingStatus(ProposalStatus.ACCEPTED)).toBeFalsy();
+        });
+    });
+});