+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingDefinitions.ts
@@ -3,11 +3,3 @@ export enum ProposalVotingTab {
     VOTES = 'VOTES',
     DETAILS = 'DETAILS',
 }
-
-export enum ProposalVotingStatus {
-    ACTIVE = 'ACTIVE',
-    PENDING = 'PENDING',
-    ACCEPTED = 'ACCEPTED',
-    REJECTED = 'REJECTED',
-    UNREACHED = 'UNREACHED',
-}